+++ lens/src/test/java/nl/bvkatwijk/lens/processor/LensCodeTest.java
@@ -1,5 +1,7 @@
 package nl.bvkatwijk.lens.processor;
 
+import nl.bvkatwijk.lens.Const;
+import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 
@@ -13,6 +15,20 @@ class LensCodeTest {
             assertEquals(
                     "\tpublic static final bLens<b> µ = new bLens<>(Lens.identity());",
                     LensCode.rootLens("b"));
+        }
+    }
+
+    @Nested
+    @DisplayName("ILens")
+    class ILensTest {
+        @Test
+        void generate() {
+            assertEquals("ILens<A, B>", LensCode.iLens("A", "B"));
+        }
+
+        @Test
+        void generate_const() {
+            assertEquals("ILens<" + Const.PARAM_SOURCE_TYPE + ", A>", LensCode.iLens("A"));
         }
     }
